requirements.txt
mail-parser
matplotlib
networkx
numpy
pandas
python-dateutil
PyYAML
scipy